PUBDEV-5742: Reverting all smbo related work as it will be done in a separate PR
