--- h2o-automl/src/main/java/ai/h2o/automl/hpsearch/SurrogateModel.java
@@ -1,7 +1,0 @@
-package ai.h2o.automl.hpsearch;
-
-import water.fvec.Frame;
-
-public abstract class SurrogateModel {
-  public abstract Frame evaluate(Frame hyperparameters, Frame train);
-}
