+++ h2o-automl/src/main/java/ai/h2o/automl/targetencoding/strategy/HPsSelectionStrategy.java
@@ -2,6 +2,5 @@ package ai.h2o.automl.targetencoding.strategy;
 
 public enum HPsSelectionStrategy {
   RGS,
-  SMBO,
   Fixed
 }
