--- h2o-automl/src/test/java/ai/h2o/automl/hpsearch/GPSurrogateModelTest.java
@@ -1,24 +1,0 @@
-package ai.h2o.automl.hpsearch;
-
-import org.junit.Test;
-
-import static org.junit.Assert.*;
-
-public class GPSurrogateModelTest {
-
-  @Test
-  public void getCovarianceMtxWithGaussianKernel() {
-  }
-
-  @Test
-  public void posteriorMeanAndVariance() {
-  }
-
-  @Test
-  public void evaluate() {
-  }
-
-  @Test
-  public void gridSearchOverGPsHyperparameters() {
-  }
-}
